+++ src/Backend/Modules/Faq/Actions/DeleteFeedback.php
@@ -11,6 +11,7 @@ namespace Backend\Modules\Faq\Actions;
 
 use Backend\Core\Engine\Base\ActionDelete as BackendBaseActionDelete;
 use Backend\Core\Engine\Model as BackendModel;
+use Backend\Form\Type\DeleteType;
 use Backend\Modules\Faq\Engine\Model as BackendFaqModel;
 
 /**
@@ -20,18 +21,33 @@ class DeleteFeedback extends BackendBaseActionDelete
 {
     public function execute(): void
     {
-        $feedbackId = $this->getRequest()->query->getInt('id');
+        $deleteForm = $this->createForm(
+            DeleteType::class,
+            null,
+            ['module' => $this->getModule(), 'action' => 'DeleteFeedback']
+        );
+        $deleteForm->handleRequest($this->getRequest());
+        if (!$deleteForm->isSubmitted() || !$deleteForm->isValid()) {
+            $this->redirect(BackendModel::createURLForAction('Index', null, null, ['error' => 'something-went-wrong']));
+
+            return;
+        }
+        $deleteFormData = $deleteForm->getData();
+
+        $feedbackId = (int) $deleteFormData['id'];
         $feedback = BackendFaqModel::getFeedback($feedbackId);
 
         // there is no feedback data, so redirect
         if (empty($feedback)) {
-            $this->redirect(BackendModel::createURLForAction('Index') . '&error=non-existing');
+            $this->redirect(BackendModel::createURLForAction('Index', null, null, ['error' => 'non-existing']));
         }
 
         BackendFaqModel::deleteFeedback($feedbackId);
-        $this->redirect(
-            BackendModel::createURLForAction('Edit') . '&amp;id=' .
-            $feedback['question_id'] . '&report=deleted#tabFeedback'
-        );
+        $this->redirect(BackendModel::createURLForAction(
+            'Edit',
+            null,
+            null,
+            ['id' => $feedback['question_id'], 'report' => 'deleted']
+        ) . '#tabFeedback');
     }
 }
